+++ CHANGELOG.md
@@ -3,9 +3,10 @@ yii2-dynamicform change Log
 
 dev-master
 ----------
-- bug: Fixed html name attribute for kartik-v/yii2-widget-colorinput
-- enh: Added support for kartik-v/yii2-widget-timepicker
-- enh: Added support for kartik-v/yii2-widget-colorinput
+- enh: Added support for "kartik-v/yii2-widget-select2"
+- bug: Fixed html name attribute for "kartik-v/yii2-widget-colorinput"
+- enh: Added support for "kartik-v/yii2-widget-timepicker"
+- enh: Added support for "kartik-v/yii2-widget-colorinput"
 
 
 version 1.0.0